refactor: 도메인의 BettingMoneyDTO 의존성 제거
- BettingMoneyDTO 내부의 도메인 로직 LottoMachine 으로 이동
- 입력 금액에 대한 예외 테스트 LottoMachineTest로 이동
- BettingMoneyDTOTest 삭제
