+++ src/main/java/lotto/domain/ticket/LottoMachine.java
@@ -2,7 +2,6 @@ package lotto.domain.ticket;
 
 import lotto.domain.ticket.ball.LottoBall;
 import lotto.domain.ticket.ball.LottoBallFactory;
-import lotto.view.dto.BettingMoneyDTO;
 import lotto.view.dto.WinLottoTicketDTO;
 
 import java.util.ArrayList;
@@ -13,16 +12,29 @@ import java.util.Set;
 import static lotto.domain.ticket.LottoTicket.LOTTO_PRICE;
 
 public abstract class LottoMachine {
+    private static final String MESSAGE_FOR_NOT_ENOUGH_MONEY = "%d는 최소 구매 금액보다 작습니다.";
 
-    public final List<LottoTicket> buyTickets(BettingMoneyDTO bettingMoney) {
-        int ticketCount = bettingMoney.getBettingMoney() / LOTTO_PRICE;
+    public final List<LottoTicket> buyTickets(int bettingMoney) {
+        validateMoney(bettingMoney);
+
+        int numberOfTickets = calculateAffordableNumber(bettingMoney);
 
         List<LottoTicket> lottoTickets = new ArrayList<>();
-        for (int i = 0; i < ticketCount; i++) {
+        for (int i = 0; i < numberOfTickets; i++) {
             lottoTickets.add(createOneTicket());
         }
 
         return lottoTickets;
+    }
+
+    private void validateMoney(int bettingMoney) {
+        if (bettingMoney < LOTTO_PRICE) {
+            throw new IllegalArgumentException(String.format(MESSAGE_FOR_NOT_ENOUGH_MONEY, bettingMoney));
+        }
+    }
+
+    private int calculateAffordableNumber(int bettingMoney) {
+        return bettingMoney / LOTTO_PRICE;
     }
 
     public abstract LottoTicket createOneTicket();
