+++ src/test/java/lotto/domain/ticket/LottoMachineTest.java
@@ -2,12 +2,12 @@ package lotto.domain.ticket;
 
 import lotto.domain.ticket.ball.LottoBall;
 import lotto.domain.ticket.ball.LottoBallFactory;
-import lotto.view.dto.BettingMoneyDTO;
 import lotto.view.dto.WinLottoTicketDTO;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.Arrays;
 import java.util.List;
@@ -15,6 +15,7 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class LottoMachineTest {
 
@@ -23,9 +24,19 @@ class LottoMachineTest {
     @CsvSource(value = {"1000,1", "1500,1", "2000,2"})
     void test1(int money, int expect) {
         LottoMachine realMachine = new AutoLottoMachine();
-        List<LottoTicket> lottoTickets = realMachine.buyTickets(new BettingMoneyDTO(money));
+        List<LottoTicket> lottoTickets = realMachine.buyTickets(money);
 
         assertThat(lottoTickets).hasSize(expect);
+    }
+
+    @DisplayName("예외 테스트: 로또 한 장 가격보다 작은 값 입력시 Exception 발생")
+    @ParameterizedTest
+    @ValueSource(ints = {-1, 0, 999})
+    void test1(int money) {
+        LottoMachine realMachine = new AutoLottoMachine();
+        assertThatThrownBy(() -> realMachine.buyTickets(money))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("%d는 최소 구매 금액보다 작습니다.", money);
     }
 
     @DisplayName("우승 로또 만들기")
@@ -51,12 +62,11 @@ class LottoMachineTest {
     void test2() {
         //given
         LottoMachine testLottoMachine = new LottoMachineForTest();
-        BettingMoneyDTO bettingMoneyDTO = new BettingMoneyDTO(1000);
 
         LottoTicket expectedTicket = new LottoTicket(aLottoBalls(1, 2, 3, 4, 5, 6));
 
         //when
-        List<LottoTicket> tickets = testLottoMachine.buyTickets(bettingMoneyDTO);
+        List<LottoTicket> tickets = testLottoMachine.buyTickets(1000);
 
         //then
         assertThat(tickets.get(0)).isEqualTo(expectedTicket);
