+++ src/test/java/lotto/domain/ticket/LottoTicketTest.java
@@ -2,7 +2,6 @@ package lotto.domain.ticket;
 
 import lotto.domain.ticket.ball.LottoBall;
 import lotto.domain.ticket.ball.LottoBallFactory;
-import lotto.view.dto.BettingMoneyDTO;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -36,7 +35,7 @@ class LottoTicketTest {
     @CsvSource(value = {"1,true", "7,false"})
     void name(int number, boolean expectedResult) {
         LottoMachine testMachine = new LottoMachineForTest();       // {1, 2, 3, 4, 5, 6} 숫자를 가지는 로또 생성기
-        LottoTicket ticket = testMachine.buyTickets(new BettingMoneyDTO(1000)).get(0);
+        LottoTicket ticket = testMachine.buyTickets(1000).get(0);
 
         LottoBall lottoBall = LottoBallFactory.getLottoBallByNumber(number);
